ASoC: Intel: atom: make sst_platform_compr_ops const
Make this const as it not modified in the file referencing it.
It only stored in the const field 'compr_ops' of a snd_soc_platform_driver
structure. Also, add const to the declaration in the header file.

Signed-off-by: Bhumika Goyal <bhumirks@gmail.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 86cb5dee237be6cf6c432299b829a5e85ed35029)

Change-Id: Idf12cfb1b6ef95a4dfce0cb6fcfd85629583f5ab
Signed-off-by: Mike Mason <michael.w.mason@intel.com>