UPSTREAM: ASoC: Intel: Skylake: Fix updown mixer module format
DSP expects length of the coefficient for updown mixer module to be 8.
So fix the max coefficient length and since we are using default values
for coefficient select which is zero, we need not explicitly initialize
it.

Signed-off-by: Guneshwor Singh <guneshwor.o.singh@intel.com>
Acked-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 3e1b96c80f1a73594b2123173c551bb0fc829b21)

BUG=b:71724897
TEST=With whole series applied + DSP Firmware and topology;
     audio playback and recording works on cnlrvp.

Change-Id: I8307ddbaa4167567da2d79e5e01cf8b57f94b7f7
Signed-off-by: Mike Mason <michael.w.mason@intel.com>