FROMGIT: ASoC: Intel: Skylake: Parse nhlt and register clock device
When NHLT endpoint is present for a SSP then we create clock for that SSP.

MCLK is consistent across endpoints and configuration for an SSP, so query
only for first endpoint for an SSP.

For SCLK/SCLKFS, the best fit is queried from the NHLT configurations which
matches the clock rate requested. Best fit is decided based on below:

1. If rate matches with multiple configurations, then the first
   configuration is selected.

2. If for a selected fs and bits_per_sample, there are multiple endpoint
   configuration match, then the configuration with max number of
   channels is selected. So, the user has to set the rate which fits
   max number of channels

So we create a platform device and pass clock information parsed as
platform data.

Signed-off-by: Sriram Periyasamy <sriramx.periyasamy@intel.com>
Signed-off-by: Jaikrishna Nemallapudi <jaikrishnax.nemallapudi@intel.com>
Signed-off-by: Subhransu S. Prusty <subhransu.s.prusty@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit bc2bd45b1f7f35b80335367f682c0ae5b2f37911
https://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git topic/intel)

BUG=b:71724897
TEST=With whole series applied + DSP Firmware and topology;
     audio playback and recording works on cnlrvp.

Change-Id: I2fb0313a82e28d9a52402bec50d0d1a6aae581f6
Signed-off-by: Mike Mason <michael.w.mason@intel.com>